backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langchain>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-groq>=0.1.5,<0.3.0
langchain-community>=0.2.0,<0.3.0
duckduckgo-search>=5.2.2
requests>=2.31.0
python-dotenv>=1.0.1
pydantic>=2.6.4
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.129.0
yfinance>=1.2.0
arxiv>=2.4.1
beautifulsoup4>=4.12.0
pytest>=8.0.0
httpx>=0.27.0
